+++ src/CommandParameters.kt
@@ -1,2 +1,2 @@
-data class CommandParameters(val argumentLine: String, val server: ChatServerState, val user: ChatUser, val room: ChatRoom, val clientID: Int)
+data class CommandParameters(val argumentLine: String, val server: ChatServerState, val user: ChatUser, val room: ChatRoom, val clientID: Long)
